adding citygraph constructor (and modifying its types)

+++ application/src/main/java/fr/delivrooom/application/model/Graphe.java
@@ -14,16 +14,16 @@ public interface Graphe {
 	 * @param j 
 	 * @return le cout de l'arc (i,j) si (i,j) est un arc ; -1 sinon
 	 */
-	public abstract int getCout(int i, int j);
+	public abstract float getCout(long i, long j);
 
 	/**
 	 * @param i 
 	 * @param j 
 	 * @return true si <code>(i,j)</code> est un arc de <code>this</code>
 	 */
-	public abstract boolean estArc(int i, int j);
+	public abstract boolean estArc(long i, long j);
 
-    public abstract List<Integer> arcs(int i);
+    public abstract List<Long> arcs(long i);
 
 
 }
